Editing tasks "in place" is working.
This is kind of fun: In the list.handlebars
we are rendering tasks/table. This renders
the template as expected and creates corresponding
view and controller in memory.

The TasksTableController is an array controller
which I configured with the "itemController" to
wrap each of it's item in to it's own
TaskTableRowController. This helps me keep state
on each row, like if we are editing that row or not.

The actions defined in table.handlebars is dispatched
to it's controller (TaskTableController). I did not
find any good way to send actions straight to the
TableRowController, so I am kinda delegating actions
to each TaskTableRow controller.

In the TaskTableRow controller I am setting up
transaction to make sure we can "roll back" if the
user clicks the Cancel button.